feat(#2426): CG stages Path/Gift/Attributes&Skills, skills move
Task 9 of the CG magic revamp: restructures the character-creation stage
sequence per Tehom's ruling that Gift sits adjacent to Path. Stage 5 is now
"Path" (path selection only), Stage 6 is "Gift" (temporarily still renders
the old MagicStage pending Task 10's GiftStage swap), and Stage 7 is
"Attributes & Skills" — skill point allocation moves out of PathStage into a
new standalone SkillsSection component mounted inside AttributesStage.

Frontend: renumbers the Stage enum and STAGE_LABELS, reorders
StageStepper's STAGES array, updates CharacterCreationPage's renderStage
dispatch, extracts SkillsSection (+ SkillPointsHeader/SkillRow/
SpecializationRow helpers) out of PathStage.tsx into its own file, and
drops the TraditionPicker mount from PathStage (tradition selection moves
into the Gift stage in Task 10 — the backend keeps the data in the
interim).

Backend: renames the Stage IntegerChoices (PATH_SKILLS->PATH,
ATTRIBUTES(6)->GIFT(6), MAGIC(7)->ATTRIBUTES(7)) and splits
get_path_skills_errors into get_path_errors (path selection only) and
get_skill_allocation_errors (skill budget, folded into get_attributes_errors)
so skill-budget errors now surface under the Attributes & Skills stage
while path-selection errors stay on Path. compute_magic_errors now maps to
Stage.GIFT. Generates migration 0009 (AlterField choices only, no data
migration) and updates stage-mapping tests in test_models.py and
test_skills_stage.py accordingly.

Docs updated in tandem: docs/systems/character_creation.md's CG Stages
table and both character-creation CLAUDE.md files (frontend + backend).
