+++ docs/systems/character_creation.md
@@ -71,9 +71,9 @@ from world.character_creation.types import (
 | 2 | Heritage | Beginnings, species, gender selected; family/tarot complete; CG points >= 0; species allowed by beginnings |
 | 3 | Lineage | Family selected, OR familyless with tarot card |
 | 4 | Distinctions | `traits_complete` flag set; CG points >= 0 |
-| 5 | Path & Skills | Path and tradition selected; skills validated against budget |
-| 6 | Attributes | All 9 stats present, valid range (10-50), multiples of 10, free points = 0 |
-| 7 | Magic | Cantrip selected; aura distribution valid; if cantrip requires facet, facet selected |
+| 5 | Path | Path selected (`get_path_errors`) |
+| 6 | Gift | Tradition, gift, >=1 technique(s), gift resonance, and Anima Check stat/skill all selected and valid (`compute_magic_errors`, 5-branch return-first gate); still renders `MagicStage` pending the #2426 Task 10 GiftStage swap |
+| 7 | Attributes & Skills | All 12 primary stats present, valid range (1-5), points remaining = 0; skill point allocation validated against budget (moved in from Path, #2426 Task 9) |
 | 8 | Appearance | Age, height band, height inches, build all set |
 | 9 | Identity | `first_name` in draft_data |
 | 10 | Final Touches | Always complete (goals are optional) |
