+++ frontend/src/character-creation/components/StageStepper.tsx
@@ -23,9 +23,9 @@ const STAGES = [
   Stage.HERITAGE,
   Stage.LINEAGE,
   Stage.DISTINCTIONS,
-  Stage.PATH_SKILLS,
+  Stage.PATH,
+  Stage.GIFT,
   Stage.ATTRIBUTES,
-  Stage.MAGIC,
   Stage.APPEARANCE,
   Stage.IDENTITY,
   Stage.FINAL_TOUCHES,
