+++ frontend/src/character-creation/components/index.ts
@@ -9,8 +9,9 @@ export { IdentityStage } from './IdentityStage';
 export { LineageStage } from './LineageStage';
 export { MagicStage } from './MagicStage';
 export { OriginStage } from './OriginStage';
-export { PathStage as PathSkillsStage } from './PathStage';
+export { PathStage } from './PathStage';
 export { ReviewStage } from './ReviewStage';
+export { SkillsSection } from './SkillsSection';
 export { StageErrorBoundary } from './StageErrorBoundary';
 export { StageStepper } from './StageStepper';
 export { StartingAreaCard } from './StartingAreaCard';
