+++ frontend/src/character-creation/types.ts
@@ -293,9 +293,9 @@ export enum Stage {
   HERITAGE = 2,
   LINEAGE = 3,
   DISTINCTIONS = 4,
-  PATH_SKILLS = 5,
-  ATTRIBUTES = 6,
-  MAGIC = 7,
+  PATH = 5,
+  GIFT = 6,
+  ATTRIBUTES = 7,
   APPEARANCE = 8,
   IDENTITY = 9,
   FINAL_TOUCHES = 10,
@@ -307,9 +307,9 @@ export const STAGE_LABELS: Record<Stage, string> = {
   [Stage.HERITAGE]: 'Heritage',
   [Stage.LINEAGE]: 'Lineage',
   [Stage.DISTINCTIONS]: 'Distinctions',
-  [Stage.PATH_SKILLS]: 'Path & Skills',
-  [Stage.ATTRIBUTES]: 'Attributes',
-  [Stage.MAGIC]: 'Magic',
+  [Stage.PATH]: 'Path',
+  [Stage.GIFT]: 'Gift',
+  [Stage.ATTRIBUTES]: 'Attributes & Skills',
   [Stage.APPEARANCE]: 'Appearance',
   [Stage.IDENTITY]: 'Identity',
   [Stage.FINAL_TOUCHES]: 'Final Touches',
