+++ src/world/character_creation/constants.py
@@ -49,9 +49,9 @@ class Stage(models.IntegerChoices):
     HERITAGE = 2, "Heritage"
     LINEAGE = 3, "Lineage"
     DISTINCTIONS = 4, "Distinctions"
-    PATH_SKILLS = 5, "Path & Skills"
-    ATTRIBUTES = 6, "Attributes"
-    MAGIC = 7, "Magic"
+    PATH = 5, "Path"
+    GIFT = 6, "Gift"
+    ATTRIBUTES = 7, "Attributes & Skills"
     APPEARANCE = 8, "Appearance"
     IDENTITY = 9, "Identity"
     FINAL_TOUCHES = 10, "Final Touches"
