+++ src/world/character_creation/tests/test_skills_stage.py
@@ -1,12 +1,20 @@
 """
-Tests for Stage 5 (Path & Skills) validation in CharacterDraft.
+Tests for skill point allocation validation in CharacterDraft.
+
+Skill allocation now lives under the Attributes & Skills stage (Stage 7);
+it used to be part of "Path & Skills" (Stage 5) before the #2426 CG stage
+restructure. ``draft.validate_path_skills()`` keeps its name (the
+``draft_data["skills"]``/``draft_data["specializations"]`` keys are
+unchanged) but its errors are now surfaced under ``Stage.ATTRIBUTES``.
 """
 
 from django.test import TestCase
 import pytest
 from rest_framework import serializers
 
+from world.character_creation.constants import Stage
 from world.character_creation.factories import CharacterDraftFactory
+from world.character_creation.validators import get_all_stage_errors
 from world.classes.factories import PathFactory
 from world.classes.models import PathStage
 from world.magic.factories import TraditionFactory
@@ -52,8 +60,7 @@ class SkillsStageValidationTests(TestCase):
             str(self.swords_spec.pk): 10,
         }
         draft.save()
-        assert draft._is_path_skills_complete() is True
-        # Also verify no exception raised
+        # No exception raised
         draft.validate_path_skills()
 
     def test_over_budget_fails(self):
@@ -67,7 +74,6 @@ class SkillsStageValidationTests(TestCase):
         draft.draft_data["skills"] = skills  # 150 points, over 110 budget
         draft.draft_data["specializations"] = {}
         draft.save()
-        assert draft._is_path_skills_complete() is False
         # Verify specific error message
         with pytest.raises(serializers.ValidationError) as exc_info:
             draft.validate_path_skills()
@@ -83,7 +89,6 @@ class SkillsStageValidationTests(TestCase):
             str(self.swords_spec.pk): 10,
         }
         draft.save()
-        assert draft._is_path_skills_complete() is False
         # Verify specific error message
         with pytest.raises(serializers.ValidationError) as exc_info:
             draft.validate_path_skills()
@@ -96,7 +101,6 @@ class SkillsStageValidationTests(TestCase):
             str(self.melee_skill.pk): 40,  # Over 30 max
         }
         draft.save()
-        assert draft._is_path_skills_complete() is False
         # Verify specific error message
         with pytest.raises(serializers.ValidationError) as exc_info:
             draft.validate_path_skills()
@@ -108,6 +112,48 @@ class SkillsStageValidationTests(TestCase):
         draft.draft_data["skills"] = {}
         draft.draft_data["specializations"] = {}
         draft.save()
-        assert draft._is_path_skills_complete() is True
-        # Also verify no exception raised
+        # No exception raised
         draft.validate_path_skills()
+
+
+class SkillStageMappingTests(TestCase):
+    """Test that skill/path validation errors surface under the correct stage (#2426)."""
+
+    @classmethod
+    def setUpTestData(cls):
+        cls.budget = SkillPointBudgetFactory(
+            path_points=50,
+            free_points=60,
+            points_per_tier=10,
+            specialization_unlock_threshold=30,
+            max_skill_value=30,
+            max_specialization_value=30,
+        )
+        cls.path = PathFactory(
+            name="Test Path for Stage Mapping",
+            stage=PathStage.PROSPECT,
+            minimum_level=1,
+        )
+        cls.tradition = TraditionFactory()
+
+    def test_missing_path_is_a_path_stage_error_independent_of_skills(self):
+        """No path selected is a Path-stage error, regardless of skill data."""
+        draft = CharacterDraftFactory(selected_path=None)
+        draft.draft_data["skills"] = {}
+        draft.save()
+        errors = get_all_stage_errors(draft)
+        assert "Select a path" in errors[Stage.PATH]
+
+    def test_skill_budget_errors_surface_under_attributes_stage(self):
+        """Over-budget skill allocation is an Attributes & Skills stage error, not a Path error."""
+        draft = CharacterDraftFactory(selected_path=self.path, selected_tradition=self.tradition)
+        skills = {}
+        for _i in range(5):
+            skill = SkillFactory()
+            skills[str(skill.pk)] = 30
+        draft.draft_data["skills"] = skills  # 150 points, over 110 budget
+        draft.draft_data["specializations"] = {}
+        draft.save()
+        errors = get_all_stage_errors(draft)
+        assert not errors[Stage.PATH]
+        assert any("exceeds budget" in e for e in errors[Stage.ATTRIBUTES])
